khushi/jobscall/requirements-prod.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
Pillow>=10.0.0
django-environ>=0.11.2
